refactor: Daily 엔티티의 date 타입을 LocalDate로 변경
- Daily의 date 필드는 시간을 사용하지 않으므로 타입을 LocalDate로 변경
- Daily에 날짜 기준 정렬을 위한 compareTo() 메서드 추가
- Strategy에서 사용하는 DoubleHandler를 global 패키지에서 삭제
- JPA Auditing을 적용하기 위해 Entity에서 직접 날짜 값을 설정하는 코드 제거
