+++ src/main/java/com/be3c/sysmetic/domain/strategy/service/DailyService.java
@@ -5,6 +5,7 @@ import com.be3c.sysmetic.domain.strategy.dto.SaveDailyRequestDto;
 import com.be3c.sysmetic.domain.strategy.dto.SaveDailyResponseDto;
 import com.be3c.sysmetic.global.common.response.PageResponse;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,5 +14,5 @@ public interface DailyService {
     SaveDailyResponseDto getIsDuplicate(Long strategyId, List<SaveDailyRequestDto> requestDtoList);
     void updateDaily(Long strategyId, Long dailyId, SaveDailyRequestDto requestDto);
     void deleteDaily(Long strategyId, Long dailyId);
-    PageResponse<DailyResponseDto> findDaily(Long strategyId, int page, LocalDateTime startDate, LocalDateTime endDate);
+    PageResponse<DailyResponseDto> findDaily(Long strategyId, int page, LocalDate startDate, LocalDate endDate);
 }
