+++ src/main/java/com/be3c/sysmetic/domain/strategy/service/MonthlyService.java
@@ -4,11 +4,11 @@ import com.be3c.sysmetic.domain.strategy.dto.MonthlyResponseDto;
 import com.be3c.sysmetic.domain.strategy.entity.Monthly;
 import com.be3c.sysmetic.global.common.response.PageResponse;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 import java.util.List;
 
 public interface MonthlyService {
-    void updateMonthly(Long strategyId, List<LocalDateTime> updatedDateList);
+    void updateMonthly(Long strategyId, List<LocalDate> updatedDateList);
     PageResponse<MonthlyResponseDto> findMonthly(Long strategyId, Integer page, Integer startYear, Integer startMonth, Integer endYear, Integer endMonth);
     Monthly calculateMonthlyData(Long strategyId, int year, int month);
 }
